refactor: separate Claude CLI discovery helpers
