+++ agentboot/claude/discovery.go
@@ -1,11 +1,8 @@
 package claude
 
 import (
-	"bufio"
 	"context"
-	"encoding/json"
 	"fmt"
-	"io"
 	"os"
 	"os/exec"
 	"path/filepath"
@@ -19,19 +16,6 @@ import (
 )
 
 const (
-	// Environment variable names
-	EnvClaudePath  = "CLAUDE_CLI_PATH"
-	EnvUseBundled  = "CLAUDE_USE_BUNDLED"
-	EnvUseGlobal   = "CLAUDE_USE_GLOBAL"
-	EnvClaudeHome  = "CLAUDE_HOME"
-	EnvNodePath    = "NODE_PATH"
-	EnvBunVersions = "BUN_VERSIONS"
-	EnvBunInstall  = "BUN_INSTALL"
-
-	// EnvBunEnv is retained for source compatibility.
-	// Deprecated: use EnvBunInstall.
-	EnvBunEnv = EnvBunInstall
-
 	claudeCLIProbeTimeout = 5 * time.Second
 )
 
@@ -108,25 +92,6 @@ func (d *CLIDiscovery) InvalidateCache() {
 	d.cachedVariant = nil
 	d.cachedEnv = nil
 	d.forceRediscover = true
-}
-
-// GetCleanEnv returns a clean environment for running Claude CLI
-func (d *CLIDiscovery) GetCleanEnv(ctx context.Context) ([]string, error) {
-	d.mu.Lock()
-	defer d.mu.Unlock()
-
-	if d.cachedEnv != nil && !d.forceRediscover {
-		return d.cachedEnv, nil
-	}
-
-	d.forceRediscover = false
-	env, err := d.buildCleanEnv(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	d.cachedEnv = env
-	return env, nil
 }
 
 // discoverCLI implements the CLI discovery logic
@@ -400,59 +365,6 @@ func (d *CLIDiscovery) runVersionProbe(ctx context.Context, path string) (string
 	return string(output), nil
 }
 
-// buildCleanEnv creates a clean environment for running Claude CLI
-func (d *CLIDiscovery) buildCleanEnv(ctx context.Context) ([]string, error) {
-	env := os.Environ()
-
-	// Remove problematic environment variables
-	cleanEnv := make([]string, 0, len(env))
-
-	for _, e := range env {
-		// Skip local node_modules paths
-		if strings.HasPrefix(e, EnvNodePath+"=") {
-			continue
-		}
-
-		// Skip Bun-specific paths that might interfere
-		if strings.HasPrefix(e, EnvBunVersions+"=") ||
-			strings.HasPrefix(e, EnvBunInstall+"=") {
-			continue
-		}
-
-		cleanEnv = append(cleanEnv, e)
-	}
-
-	// Ensure PATH doesn't contain local node_modules
-	cleanEnv = d.cleanPATH(cleanEnv)
-
-	return cleanEnv, nil
-}
-
-// cleanPATH removes local node_modules directories from PATH
-func (d *CLIDiscovery) cleanPATH(env []string) []string {
-	for i, e := range env {
-		if strings.HasPrefix(e, "PATH=") {
-			pathValue := e[5:]
-			paths := strings.Split(pathValue, string(os.PathListSeparator))
-
-			cleanPaths := make([]string, 0, len(paths))
-			for _, p := range paths {
-				// Skip node_modules paths
-				if strings.Contains(p, "node_modules") ||
-					strings.Contains(p, ".bun") {
-					continue
-				}
-				cleanPaths = append(cleanPaths, p)
-			}
-
-			env[i] = "PATH=" + strings.Join(cleanPaths, string(os.PathListSeparator))
-			break
-		}
-	}
-
-	return env
-}
-
 // parseVersion extracts version from CLI output
 func parseVersion(output string) string {
 	lines := strings.Split(strings.TrimSpace(output), "\n")
@@ -529,177 +441,7 @@ func FindClaudeCLI(ctx context.Context) (*CLIVariant, error) {
 	return defaultDiscovery.FindClaudeCLI(ctx)
 }
 
-// GetCleanEnv is a convenience function for getting clean environment
-func GetCleanEnv(ctx context.Context) ([]string, error) {
-	return defaultDiscovery.GetCleanEnv(ctx)
-}
-
 // InvalidateDiscoveryCache invalidates the global discovery cache
 func InvalidateDiscoveryCache() {
 	defaultDiscovery.InvalidateCache()
-}
-
-// StreamToStdin streams messages to the stdin of a running process
-func StreamToStdin(ctx context.Context, stdin io.WriteCloser, messages <-chan any) error {
-	logrus.Debugln("[StreamToStdin] Starting to stream messages to stdin")
-
-	// Use buffered writer for efficient I/O and ensure data is flushed
-	writer := bufio.NewWriter(stdin)
-
-	encoder := json.NewEncoder(writer)
-
-	messageCount := 0
-	for {
-		select {
-		case <-ctx.Done():
-			logrus.Debugln("[StreamToStdin] Context cancelled, stopping")
-			return ctx.Err()
-		case msg, ok := <-messages:
-			if !ok {
-				// Channel closed, flush any remaining data and return
-				// IMPORTANT: Do NOT close stdin here - it will be managed by the Query
-				writer.Flush()
-				logrus.Debugf("[StreamToStdin] Message channel closed after sending %d messages", messageCount)
-				return nil
-			}
-
-			messageCount++
-			logrus.Debugf("[StreamToStdin] Sending message #%d", messageCount)
-
-			if err := encoder.Encode(msg); err != nil {
-				return fmt.Errorf("encode message: %w", err)
-			}
-			// Flush immediately after each message to ensure prompt delivery
-			if err := writer.Flush(); err != nil {
-				return fmt.Errorf("flush message: %w", err)
-			}
-		}
-	}
-}
-
-// StreamReader reads line-delimited JSON from a reader
-type StreamReader struct {
-	scanner *bufio.Scanner
-}
-
-// NewStreamReader creates a new stream reader
-func NewStreamReader(r io.Reader) *StreamReader {
-	return &StreamReader{
-		scanner: bufio.NewScanner(r),
-	}
-}
-
-// Next reads the next JSON object from the stream
-func (r *StreamReader) Next() (map[string]interface{}, error) {
-	if !r.scanner.Scan() {
-		if err := r.scanner.Err(); err != nil {
-			return nil, err
-		}
-		return nil, io.EOF
-	}
-
-	line := r.scanner.Bytes()
-	var data map[string]interface{}
-	if err := json.Unmarshal(line, &data); err != nil {
-		return nil, fmt.Errorf("decode json: %w", err)
-	}
-
-	return data, nil
-}
-
-// ReadAll reads all remaining objects from the stream
-func (r *StreamReader) ReadAll() ([]map[string]interface{}, error) {
-	var results []map[string]interface{}
-
-	for {
-		data, err := r.Next()
-		if err == io.EOF {
-			break
-		}
-		if err != nil {
-			return nil, err
-		}
-		results = append(results, data)
-	}
-
-	return results, nil
-}
-
-// StreamWriter writes line-delimited JSON to a writer
-type StreamWriter struct {
-	writer io.Writer
-	closed bool
-	mu     sync.Mutex
-}
-
-// NewStreamWriter creates a new stream writer
-func NewStreamWriter(w io.Writer) *StreamWriter {
-	return &StreamWriter{
-		writer: w,
-	}
-}
-
-// Write writes a JSON object to the stream
-func (w *StreamWriter) Write(data map[string]interface{}) error {
-	w.mu.Lock()
-	defer w.mu.Unlock()
-
-	if w.closed {
-		return fmt.Errorf("stream writer is closed")
-	}
-
-	// Encode with newline
-	buf, err := json.Marshal(data)
-	if err != nil {
-		return err
-	}
-
-	buf = append(buf, '\n')
-	_, err = w.writer.Write(buf)
-	return err
-}
-
-// Close closes the stream writer
-func (w *StreamWriter) Close() error {
-	w.mu.Lock()
-	defer w.mu.Unlock()
-
-	w.closed = true
-	if closer, ok := w.writer.(io.Closer); ok {
-		return closer.Close()
-	}
-	return nil
-}
-
-// FormatEnv formats an environment variable as KEY=VALUE
-func FormatEnv(key, value string) string {
-	return fmt.Sprintf("%s=%s", key, value)
-}
-
-// MergeEnv merges custom environment variables with base environment
-// Custom variables override base ones with the same key
-func MergeEnv(base []string, custom []string) []string {
-	result := make([]string, 0, len(base)+len(custom))
-	baseMap := make(map[string]string)
-
-	// Build map from base
-	for _, e := range base {
-		if idx := strings.IndexByte(e, '='); idx > 0 {
-			baseMap[e[:idx]] = e[idx+1:]
-		} else {
-			result = append(result, e)
-		}
-	}
-
-	// Add base env (will be overridden by custom if key matches)
-	for k, v := range baseMap {
-		result = append(result, FormatEnv(k, v))
-	}
-
-	// Add custom env (overrides)
-	for _, e := range custom {
-		result = append(result, e)
-	}
-
-	return result
 }
